x86_64: Introduce an I/O bitmap struct
When virtualizing a guest, VT-x provides a way to describe which I/O ports
will be passed through and which will cause an exit when accessed. This
information is managed by a bitmap that will be set into the VMCS area.
Introduce an IO bitmap struct that abstracts this bitmap.

Signed-off-by: KarimAllah Ahmed <karim.allah.ahmed@gmail.com>